Dashboard/requirements.txt
py-cord
PyNaCl
pycord-ext-ipc
Quart-Discord-any[pycord]